Add `points` prop to Waveform, refactor helpers
Previously, the `Waveform` component only took a `shape` string.

I realized that when it comes time to handle waveform addition,
I'll need to be able to supply arbitrary points to Waveform, since it
won't always be a neat basic waveform.

Doing so required refactoring the waveform helpers to separate the
mathematical "get sine values from -1 to 1" bit from the "create an
SVG path I can actually draw" bit. Ultimately this was a good refactor,
even aside from this change.